IPO Status added

+++ src/main/java/com/virtu_stock/Subscription/Subscription.java
@@ -16,5 +16,5 @@ import lombok.NoArgsConstructor;
 public class Subscription {
     private String name;
     @Column(name = "subscription_value")
-    private Integer subsvalue;
+    private Double subsvalue;
 }
